feat: Add lexographical palindrome checker
This commit adds a new class `lexographic_pal` to the `Lexographical_palindrome` package. The class contains a `main` method that checks if a given string can be rearranged to form a palindrome. The algorithm counts the frequency of each character in the string and checks if it is possible to form a palindrome by rearranging the characters. If it is possible, the program prints "Yes".

The code is currently incomplete and contains some syntax errors. The missing semicolon after `int freq[] = new int[max]` and the empty string assignment `String odd_String=''` need to be fixed.

This feature is being added to enhance the functionality of the application and provide a way to check if a string can be rearranged to form a palindrome.